+++ include/beast/http/detail/type_traits.hpp
@@ -56,12 +56,14 @@ protected:
     string_view get_method_impl() const;
     string_view get_target_impl() const;
     string_view get_reason_impl() const;
-    bool get_keep_alive_impl(bool) const;
+    bool get_chunked_impl() const;
+    bool get_keep_alive_impl(unsigned) const;
     void set_method_impl(string_view);
     void set_target_impl(string_view);
     void set_reason_impl(string_view);
+    void set_chunked_impl(bool);
+    void set_content_length_impl(boost::optional<std::uint64_t>);
     void set_keep_alive_impl(unsigned, bool);
-    void prepare_payload_impl(bool, boost::optional<std::uint64_t>);
 };
 
 template<class T, class = beast::detail::void_t<>>
@@ -114,56 +116,69 @@ struct is_fields_helper : T
 
     template<class U = is_fields_helper>
     static auto f4(int) -> decltype(
-        std::declval<bool&>() = std::declval<U const&>().get_keep_alive_impl(
-            std::declval<unsigned>()),
+        std::declval<bool&>() = std::declval<U const&>().get_chunked_impl(),
         std::true_type());
     static auto f4(...) -> std::false_type;
     using t4 = decltype(f4(0));
 
     template<class U = is_fields_helper>
     static auto f5(int) -> decltype(
-        void(std::declval<U&>().set_method_impl(std::declval<string_view>())),
+        std::declval<bool&>() = std::declval<U const&>().get_keep_alive_impl(
+            std::declval<unsigned>()),
         std::true_type());
     static auto f5(...) -> std::false_type;
     using t5 = decltype(f5(0));
 
     template<class U = is_fields_helper>
     static auto f6(int) -> decltype(
-        void(std::declval<U&>().set_target_impl(std::declval<string_view>())),
+        void(std::declval<U&>().set_method_impl(std::declval<string_view>())),
         std::true_type());
     static auto f6(...) -> std::false_type;
     using t6 = decltype(f6(0));
 
     template<class U = is_fields_helper>
     static auto f7(int) -> decltype(
-        void(std::declval<U&>().set_reason_impl(std::declval<string_view>())),
+        void(std::declval<U&>().set_target_impl(std::declval<string_view>())),
         std::true_type());
     static auto f7(...) -> std::false_type;
     using t7 = decltype(f7(0));
 
     template<class U = is_fields_helper>
     static auto f8(int) -> decltype(
-        void(std::declval<U&>().set_keep_alive_impl(
-            std::declval<unsigned>(),
-            std::declval<bool>()
-        )), std::true_type());
+        void(std::declval<U&>().set_reason_impl(std::declval<string_view>())),
+        std::true_type());
     static auto f8(...) -> std::false_type;
     using t8 = decltype(f8(0));
 
     template<class U = is_fields_helper>
     static auto f9(int) -> decltype(
-        void(std::declval<U&>().prepare_payload_impl(
-            std::declval<bool>(),
-            std::declval<boost::optional<std::uint64_t>>())),
+        void(std::declval<U&>().set_chunked_impl(std::declval<bool>())),
         std::true_type());
     static auto f9(...) -> std::false_type;
     using t9 = decltype(f9(0));
 
+    template<class U = is_fields_helper>
+    static auto f10(int) -> decltype(
+        void(std::declval<U&>().set_content_length_impl(
+            std::declval<boost::optional<std::uint64_t>>())),
+        std::true_type());
+    static auto f10(...) -> std::false_type;
+    using t10 = decltype(f10(0));
+
+    template<class U = is_fields_helper>
+    static auto f11(int) -> decltype(
+        void(std::declval<U&>().set_keep_alive_impl(
+            std::declval<unsigned>(),
+            std::declval<bool>())),
+        std::true_type());
+    static auto f11(...) -> std::false_type;
+    using t11 = decltype(f11(0));
+
     using type = std::integral_constant<bool,
-        t1::value && t2::value && t3::value &&
-        t4::value && t5::value && t6::value &&
-        t7::value && t8::value && t9::value
-    >;
+         t1::value &&  t2::value && t3::value &&
+         t4::value &&  t5::value && t6::value &&
+         t7::value &&  t8::value && t9::value &&
+        t10::value && t11::value>;
 };
 
 } // detail